backend/requirements.txt
Django
djangorestframework
psycopg2-binary
tenacity
aiohttp
pytest
pytest-django
beautifulsoup4
pytest-asyncio
